Fix 64 bit issues in QIODevicePrivateLinearBuffer
The API was using int, not qint64 leading to implicit
truncation of numbers in a few places

Task-number: QTBUG-40974

+++ src/corelib/io/qiodevice_p.h
@@ -84,13 +84,13 @@ public:
         first = buf;
         capacity = 0;
     }
-    int size() const {
+    qint64 size() const {
         return len;
     }
     bool isEmpty() const {
         return len == 0;
     }
-    void skip(int n) {
+    void skip(qint64 n) {
         if (n >= len) {
             clear();
         } else {
@@ -106,14 +106,14 @@ public:
         first++;
         return ch;
     }
-    int read(char* target, int size) {
+    int read(char* target, qint64 size) {
         int r = qMin(size, len);
         memcpy(target, first, r);
         len -= r;
         first += r;
         return r;
     }
-    int peek(char* target, int size) {
+    int peek(char* target, qint64 size) {
         int r = qMin(size, len);
         memcpy(target, first, r);
         return r;
@@ -124,7 +124,7 @@ public:
         len += size;
         return writePtr;
     }
-    void chop(int size) {
+    void chop(qint64 size) {
         if (size >= len) {
             clear();
         } else {
@@ -136,7 +136,7 @@ public:
         clear();
         return retVal;
     }
-    int readLine(char* target, int size) {
+    int readLine(char* target, qint64 size) {
         int r = qMin(size, len);
         char* eol = static_cast<char*>(memchr(first, '\n', r));
         if (eol)
@@ -158,7 +158,7 @@ public:
         len++;
         *first = c;
     }
-    void ungetBlock(const char* block, int size) {
+    void ungetBlock(const char* block, qint64 size) {
         if ((first - buf) < size) {
             // underflow, the existing valid data needs to move to the end of the (potentially bigger) buffer
             makeSpace(len + size, freeSpaceAtStart);
@@ -191,7 +191,7 @@ private:
 
 private:
     // length of the unread data
-    int len;
+    qint64 len;
     // start of the unread data
     char* first;
     // the allocated buffer
